Avoid using Dehacked for customizing/colorizing strings
This replaces any call to DEH_AddStringReplacement() with alternative
functions to colorize (or modify in the braoder sense) in-game strings
on-the-fly. This was meant as an attempt to get rid of the DEH
checksum mismatch warnings when connecting mixed Crispy and Choco
server and clients for network games.

However, this turned out to be a red herring. Even when the Dehacking
of strings was removed, the warning would persist. Why? Because there
are two states modified in info.c, both affecting the duration and
brightness of the SSG muzzle flash states. And even if I removed these
modifications, the DEH checksums would still differ. This is, because
many mobjs have been modified to carry a flag that results in them
being rendered translucently.

I don't know if the warning would disappear if I reverted all of these
changes, and I am not going to check this out. Instead, I will drop
the idea of being DEH-clean with Choco right here.

Why? First, I think we have to accept the fact that Crispy Doom by now
has deviated quite a bit from Choco, even "under the hood". As long as
these changes remain "harmless" (and I promise that, everything else
would be a severe bug) I see no reason to change this. Second, I could
try to hide these changes and apply them at runtime, just as I tried
with the string colorizations. But, on the one hand, this would result
in ugly code and workarounds and applying all these modifications at
runtime will result in performance penalties. On the other hand, the
changes are there, so what! Why not admit that we are not DEH-clean
with Choco anymore? Yes, Crispy has changed some strings, some states
and some mobjs. But, again, I promise these changes remain uncritical,
that's one of the design goals of Crispy Doom. The warning is right
there to raise the players' attentions to the differences in two
engines that are going to connect over network, so why not leave it
just in place?

Hugh, this fixes #105